+++ src/RcppExports.cpp
@@ -1476,6 +1476,18 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// foceiOuterH
+NumericMatrix foceiOuterH(NumericVector theta, double relStep);
+RcppExport SEXP _nlmixr2est_foceiOuterH(SEXP thetaSEXP, SEXP relStepSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< NumericVector >::type theta(thetaSEXP);
+    Rcpp::traits::input_parameter< double >::type relStep(relStepSEXP);
+    rcpp_result_gen = Rcpp::wrap(foceiOuterH(theta, relStep));
+    return rcpp_result_gen;
+END_RCPP
+}
 // nmNearPD_
 RObject nmNearPD_(RObject x, bool keepDiag, bool do2eigen, bool doDykstra, bool only_values, double eig_tol, double conv_tol, double posd_tol, int maxit, bool trace);
 RcppExport SEXP _nlmixr2est_nmNearPD_(SEXP xSEXP, SEXP keepDiagSEXP, SEXP do2eigenSEXP, SEXP doDykstraSEXP, SEXP only_valuesSEXP, SEXP eig_tolSEXP, SEXP conv_tolSEXP, SEXP posd_tolSEXP, SEXP maxitSEXP, SEXP traceSEXP) {