+++ tests/testthat.R
@@ -102,7 +102,8 @@ if (identical(Sys.info()[["sysname"]], "Darwin")) {
   c(
     "focei-wang2007-boxcox-lnorm", "nlme-cov", "agq-fast-grad",
     "focei-ll-fast-grad-fit", "focei-fast-methods-fit", "odeswap-fit",
-    "focei-factr-fit", "focei-reproducible", "focei-hessian-etastep"
+    "focei-factr-fit", "focei-reproducible", "focei-hessian-etastep",
+    "focei-outer-hessian-fit", "focei-outer-hessian-nlminb-fit", "focei-outer-hessian-family"
   ),
   # batch 6
   c(